+++ packages/lumx-icons/override/override-icons/config.js
@@ -1,4 +1,9 @@
 module.exports = {
+    'x-twitter': {
+        replace: true,
+        cssCode: '\\F0544',
+    },
+    // Icon restored from MDI v4 to avoid breaking change
     'google-pages': {
         cssCode: '\\F2BA',
     },